fix incorrect path comparison
Currently, when checking whether we are in the wiki or not the
comparison will only return true if we are in the root of the wiki,
since the comparison is exact folder to folder. Check instead if the
path of the wiki is contained in the path of the current file. Move the
comments to the appropriate path of the conditional.