fix: backlink issue when creating zettels with 'z'
Previously the backlinks didn't work as the code was abusing the third
argument of vimwiki#base#open_link. The trick is that vimwiki accepts
"absolute paths" in the sense that it is absolute from the root of the
wiki. Use this feature to create the zettels in the right place.