new selected: create proper link name
After adding the relative path for the zettelkasten creating links from
the selection properly created the zettel in the designated folder, but
created the link as if it was in the current folder. Separate the
behaviour into two parts. If we are within the zettelkasten, create the
link pointing to the current folder (this will fail if the zettelkasten
is not flat, but current designs implicitly assume flatness elsewhere as
well). If we are not in the zettelkasten, make the link as a (vimwiki)
absolute path to the new zettel and prefix the link name with the
zettelkasten path as well.