make link paths relative to current file
When editing file a/b/c.md and trying to link to a/f/d.md, the current
behaviour is to link from the root of the wiki, which doesn't work.
I.e. the link would be 'a/f/d.md', when we need '../f/d.md'.

So here we make the link relative to the file we're editing.